fix: shift orders onDelete task or column

+++ controllers/columnController.js
@@ -1,3 +1,4 @@
+const { Op } = require('sequelize');
 const errorHandler = require('express-async-handler');
 const { Column } = require('../models');
 
@@ -39,29 +40,45 @@ const getColumnById = errorHandler(async (req, res) => {
 
 const updateColumnByID = errorHandler(async (req, res) => {
   const { id } = req.params;
-  const { title, order } = req.body;
+  const { title } = req.body;
 
-  if (!title && !id) {
+  if (!title || !id) {
     res.status(400);
-    throw new Error('Column must contain title or id');
+    throw new Error('Column must contain title and id');
   }
 
-  await Column.update({ title, order }, { where: { id } });
+  await Column.update({ title }, { where: { id } });
   res.status(200);
   return res.json({ message: 'Column has been updated' });
 });
 
 const deleteColumnById = errorHandler(async (req, res) => {
   const { id } = req.params;
+  const column = await Column.findOne({ where: { id } });
+
+  if (!column) {
+    res.status(404);
+    throw new Error('Column not found');
+  }
+
   const destroyed = await Column.destroy({ where: { id } });
 
   if (destroyed === 1) {
+    const { BoardId, order } = column.dataValues;
+    const cols = await Column.findAll({
+      where: {
+        BoardId,
+        order: { [Op.gt]: order },
+      },
+    });
+    const newOrders = cols.map((col) => col?.dataValues.order - 1);
+    cols.forEach((col, i) => col.update({ order: newOrders[i] }));
     res.status(200);
     return res.json({ message: 'Column has been deleted' });
   }
 
-  res.status(404);
-  throw new Error('Column not found');
+  res.status(503);
+  throw new Error('Column is not deleted. Try again');
 });
 
 module.exports = {
